CRM_Automation/requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
pyairtable>=2.3.0
pyyaml>=6.0.0
